Divide test_ProcessorStage.py per how-to: tests.
- Move common function to separate file lib.py.
